# project_configuration/JsonLd/Generator/Factory_Smart_Data_Model/MultiFiles/Working/Factory_3/WorkingYaml/openapi.yaml
components:
  schemas:
...
    Sensor:
      type: object
      required:
        - id
        - name
        - description
        - type
        - state
      properties:
        id:
          type: string
          description: Unique identifier for the sensor.
        name:
          type: string
          description: Name of the sensor.
        description:
          type: string
          description: Detailed description of the sensor.
        type:
          type: string
          description: Type of the sensor (e.g., "position", "weight", etc.).
        state:
          type: string
          enum:
            - active
            - inactive
          description: State of the sensor (active or inactive).
paths:
...
  /sensors:
    get:
      summary: Get a list of sensors.
      responses:
        "200":
          description: A list of sensors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sensor"
    post:
      summary: Create a new sensor.
      requestBody:
        description: Create a new sensor.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sensor"
      responses:
        "201":
          description: Sensor successfully created.
    put:
      summary: Update an existing sensor.
      requestBody:
        description: Update an existing sensor.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sensor"
      responses:
        "200":
          description: Sensor successfully updated.
    patch:
      summary: Update a sensor dynamically.
      requestBody:
        description: Update a sensor dynamically based on specific data or context changes.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sensor"
      responses:
        "200":
          description: Sensor successfully updated.
    delete:
      summary: Delete a sensor.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Sensor successfully deleted.
info:
  title: Smart Data Model for Factory
  version: 1.0.0
  description: Data model for factories, with sectors, containers, and industry type.
openapi: "3.0.0"